requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
begins==0.9
blinker==1.4
Bottleneck==1.3.2
brotlipy==0.7.0
build==0.9.0
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
cloudpickle==2.0.0
colorama==0.4.4
cryptography==36.0.0
cycler==0.11.0
Cython==0.29.25
cytoolz==0.11.0
dask==2022.5.0
dataclasses==0.8
DPA==0.0.3
easycython==1.0.7
et-xmlfile==1.1.0
fastcluster==1.2.6
fonttools==4.28.3
frozenlist==1.2.0
fsspec==2022.3.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==2.10.0
idna==3.1
imageio==2.9.0
importlib-metadata==4.8.2
joblib==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
llvmlite==0.38.1
locket==1.0.0
Markdown==3.3.6
matplotlib==3.4.3
mkl-service==2.4.0
multidict==5.2.0
munkres==1.1.4
natsort==7.0.1
networkx==2.6.3
nibabel==3.2.1
nitime==0.9
numba==0.55.2
numexpr==2.8.1
numpy==1.21.4
oauthlib==3.1.1
olefile==0.46
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
partd==1.2.0
pep517==0.13.0
Pillow==8.4.0
pip==21.2.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydicom==2.2.2
pydot==1.4.1
PyJWT==2.3.0
pynrrd==0.4.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
PyQt5_sip==4.19.18
pyreadline==2.1
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
pyu2f==0.1.5
PyWavelets==1.3.0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.7.3
seaborn==0.11.2
setuptools==58.0.4
sip==4.19.24
six==1.16.0
sonpy==1.8.5
tensorboard==2.7.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.10.1
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tslearn==0.5.2
typing_extensions==4.0.1
unicodedata2==13.0.0.post2
urllib3==1.26.7
Werkzeug==2.0.1
wheel==0.37.0
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0